Update token expiration types and add logout endpoint
Switched token expiration from i64/long to i32/int in both backend (Rust) and frontend (C#) for consistency. Added a logout endpoint in the Rust backend to revoke refresh tokens. Implemented LogoutAsync in the frontend AuthApiService. Updated DTOs and AuthService to use new expiration types. Improved UI text and imports, and refactored code for clarity.

+++ TabajarasInterview.Web/DTOs/LoginResponse.cs
@@ -6,7 +6,7 @@ namespace TabajarasInterview.Web.DTOs;
 public sealed class LoginResponse
 {
     public string AccessToken { get; set; } = string.Empty;
-    public long ExpiresIn { get; set; }
+    public int ExpiresIn { get; set; }
     public string RefreshToken { get; set; } = string.Empty;
     public UserResponse User { get; set; } = new();
 }
